@@ -1,11 +1,12 @@
 {
-  "name": "codeception/visualception",
+  "name": "paulmskim/module-visualception",
   "description": "Visual regression tests for Codeception",
-  "version": "1.0.4",
-  "homepage": "https://github.com/Codeception/VisualCeption",
+  "version": "2.0.0",
+  "type": "library",
+  "homepage": "https://github.com/paulmskim/module-visualception",
   "license": "Apache-2.0",
   "require": {
-    "php": ">=5.4.0",
+    "php": ">=5.6.0",
     "ext-imagick": "*"
   },
   "require-dev": {
@@ -13,6 +14,7 @@
     "phpunit/phpunit": "6.5.14"
   },
   "autoload": {
-    "classmap": ["module/"]
-  }
+    "classmap": ["src/"]
+  },
+  "minimum-stability": "dev"
 }
